refactor: rename to rql
